requirements.txt
fastapi
uvicorn
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn
python-dotenv
openai
requests
pydantic